Modularize gftools add-font
